refactor(inter): refactor class InterAtomList (atom/inter_atom.h/.cpp)
1. combine xinter,vinter,finter,dfinter,..., into list<AtomElements>; 2. adapter refactor codes in atom.cpp/atom.h,atom_dump.cpp and newton_motion.cpp (other file was not changed); 3. move inter atoms pack and unpack codes from namespace pack to class InterAtomList (but the code did not get change).

+++ src/atom.h
@@ -53,7 +53,7 @@ public :
 
     void getIntertosend(int d, int direction, double ghostlengh, std::vector<int> &sendlist);
 
-    int getintersendnum(int dimension, int direction);
+    unsigned long getintersendnum(int dimension, int direction);
 
     /**
      * set velocity of a atom whose position is specified by array @param lat
